move manual info to config as well
